requirements.txt
web3==6.15.0
eth-account==0.11.0
aiohttp==3.9.3
asyncio==3.4.3
redis==5.0.3
python-dotenv==1.0.1
cdp-sdk>=0.14.0
coinbase-advanced-py==1.4.1
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
websockets==12.0
textblob==0.18.0
tweepy==4.14.0
loguru==0.7.2
pytest==8.1.1
pytest-asyncio==0.23.5
